backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain>=0.1.0
langchain-groq>=0.1.0
tavily-python>=0.3.0
python-docx>=1.1.0
fpdf2>=2.7.6
fpdf2>=2.7.6
pydantic>=2.5.0
langchain-google-genai>=0.0.9
python-dotenv>=1.0.0
supabase


